administrator modules updates

+++ Models/global/ListUserModel.cs
@@ -2,7 +2,7 @@ namespace itsppisapi.Models
 {
     public class ListUserModel
     {
-        public int USER_ID { get; set; }
+        public decimal USER_ID { get; set; }
         public string USER_NAME { get; set; }
 
     }
